Getting stuff set up
Sorry, had to delete preferences.xml, Picker.java, and
PickerPreferences.java to make aapt happy. Hope we weren't
planning on using them in the future; if we were, we can
figure out a way to add them back in.

Deleted .project because typically IDE-generated files
don't need to be checked in (a needless source of conflicts)